transcribe/requirements.txt
fastapi
pydantic
python-multipart
hypercorn[trio]
Werkzeug
openai-whisper
